All search related methods moved to a new created class (AcademicSearchService) (note: I can not give the name as SearchService because there are already two different implemented things use the same name so I just want to keeps thing simplified)
