+++ src/main/java/io/academic/service/OaiService.java
@@ -68,6 +68,10 @@ public class OaiService {
     private DcParseService dcParseService;
 
 
+    public RestHighLevelClient getRestClient() {
+        return restClient;
+    }
+
     RestHighLevelClient restClient = new RestHighLevelClient( RestClient.builder(
             new HttpHost("localhost", 9200, "http"),
             new HttpHost("localhost", 9201, "http")));
@@ -174,87 +178,7 @@ public class OaiService {
 //        return LocalDateTime.parse(string, formatter);
     }
 
-
-
-    public String search(String q) throws IOException {
-
-        SearchRequest searchRequest = new SearchRequest("harvester");
-        SearchSourceBuilder searchSourceBuilder = new SearchSourceBuilder();
-        searchSourceBuilder.query(QueryBuilders.termQuery("dc",q));
-        searchSourceBuilder.sort(new FieldSortBuilder("title.keyword").order(SortOrder.DESC));
-        searchSourceBuilder.fetchSource("title","");
-        searchRequest.source(searchSourceBuilder);
-        SearchResponse searchResponse = restClient.search(searchRequest);
-        String result = searchResponse.toString();
-
-        Gson gson = new GsonBuilder().setPrettyPrinting().create();
-        JsonParser jp = new JsonParser();
-        JsonElement je = jp.parse(result);
-        String prettyJsonString = gson.toJson(je);
-
-        System.out.println(prettyJsonString);
-
-        return prettyJsonString; //pre tag for json, otherwise it didnt show pretty in browser
-
-    }
-
-    public String searchPretty(String q) throws IOException {
-
-
-        return "<pre>"+search(q)+"</pre>"; //pre tag for json, otherwise it didnt show pretty in browser
-
-    }
-
-    public String searchForm(String q) throws IOException {
-
-        return search(q); //pre tag for json, otherwise it didnt show pretty in browser
-
-    }
-
-    public String getAll() throws IOException {
-
-
-        final Scroll scroll = new Scroll(TimeValue.timeValueMinutes(1L));
-
-        SearchRequest searchRequest = new SearchRequest("harvester");
-        searchRequest.scroll(scroll);
-        SearchSourceBuilder searchSourceBuilder = new SearchSourceBuilder();
-        searchSourceBuilder.query(QueryBuilders.matchAllQuery());
-        searchRequest.source(searchSourceBuilder);
-
-        SearchResponse searchResponse = restClient.search(searchRequest);
-        String scrollId = searchResponse.getScrollId();
-        SearchHit[] searchHits = searchResponse.getHits().getHits();
-        String result="";
-        Gson gson = new GsonBuilder().setPrettyPrinting().create();
-        JsonParser jp = new JsonParser();
-        JsonElement je ;
-        String prettyJsonString ;
-        while (searchHits != null && searchHits.length > 0) {
-            SearchScrollRequest scrollRequest = new SearchScrollRequest(scrollId);
-            scrollRequest.scroll(scroll);
-            searchResponse = restClient.searchScroll(scrollRequest);
-            scrollId = searchResponse.getScrollId();
-            searchHits = searchResponse.getHits().getHits();
-
-
-            je = jp.parse(searchResponse.toString());
-            prettyJsonString = gson.toJson(je);
-            result+=prettyJsonString;
-
-
-
-        }
-
-        ClearScrollRequest clearScrollRequest = new ClearScrollRequest();
-        clearScrollRequest.addScrollId(scrollId);
-        ClearScrollResponse clearScrollResponse = restClient.clearScroll(clearScrollRequest);
-        boolean succeeded = clearScrollResponse.isSucceeded();
-
-
-        return "<pre>"+result+"</pre>"; //pre tag for json, otherwise it didnt show pretty in browser
-    }
-
+    
     public void delete() throws IOException {
 
         //TODO:check if there is any indices with that name
